CASMPET-7604 CRD install/upgrade fix
Helm only installs CRDs during an initial install,
and it does not install CRDs that already exist.
During an upgrade helm does not update CRDs, then
need to be done manually.

We have modified the CRDs in the /files directory,
but those only get applied manually during the upgrade
process, on an initial install they do not get put in
place, the CRDs in the postgress-operator subchart get
installed.

With moving the modified CRDs from the /files directory to
the /crds directory, helm will install the CRDs in that
top level /crds directory instead of the CRDs in the
sub-chart.  That resolves the issue with an upgrade and
an initial install installing different CRDs.  This
change means the code in the upgrade prerequisits.sh
script needs to be modified to find the CRDs in the
/crds directory instead of the /files directory.

This PR also updates the chart version 1.10.1 -> 1.10.2,
previous changes neglected to change the chart version,
so there has been confusion as to what is in the 1.10.1
chart.